# grov-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grōv Restoration Platform API",
        "description": "Production-quality REST API backend for Grōv field restoration tracking, monitoring, community hub, and leaderboard.",
        "contact": {
            "name": "Grōv Engineering Team",
            "email": "support@grov.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Grōv Primary API Server v1"
        }
    ],
    "paths": {
...
        "/user/profile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get current user profile",
                "description": "Returns profile details, statistics, role, and selected interests.",
                "operationId": "fe3a194a4e23e01c363b58cb0491f778",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update profile details",
                "description": "Updates name, location, bio, or role.",
                "operationId": "5be47c3f2bb37025d2bfe7754e6f2ee9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "bio": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "volunteer",
                                            "coordinator"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/avatar": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Upload profile avatar photo",
                "description": "Stores avatar image file and updates user avatar_path.",
                "operationId": "93c05b6d6ba56d027806ca7563e9014b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar uploaded successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/interests": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "List all interest tags & user selected interests",
                "operationId": "3d6abcf0f3de91523d6f631c532c827e",
                "responses": {
                    "200": {
                        "description": "Interests retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update user interest tag selection",
                "operationId": "329bd590ca394c063857aa06750c2a4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "interest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interests updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/settings": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update notification and application settings",
                "operationId": "61971c4e67f43f9401386b0deb72da13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "settings": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/account": {
            "delete": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Delete user account",
                "description": "Soft deletes user account.",
                "operationId": "f3430a8f77e34b69409b475029986cc5",
                "responses": {
                    "200": {
                        "description": "Account deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum bearer token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}